+++ src/app/modules/admin/components/gigs/gigs.component.ts
@@ -3,6 +3,7 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { FrontendService } from 'src/app/services/frontend.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-gigs',
@@ -37,12 +38,43 @@ getAllgigs(){
 }
 
 deleteGig(id:string){
-  this.service.deleteGig(id).subscribe((response)=>{
-    console.log(response)
-    if(response.deleteGig===true){
-      this.ngOnInit()
+
+  Swal.fire({
+    title: 'Do you want to delete the gig?',
+    
+    showCancelButton: true,
+    confirmButtonText: 'Delete',
+    
+  }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+    if (result.isConfirmed) {
+      this.service.deleteGig(id).subscribe((response)=>{
+        const Toast = Swal.mixin({
+          toast: true,
+          position: 'top-end',
+          showConfirmButton: false,
+          timer: 3000,
+          timerProgressBar: true,
+          didOpen: (toast) => {
+            toast.addEventListener('mouseenter', Swal.stopTimer)
+            toast.addEventListener('mouseleave', Swal.resumeTimer)
+          }
+        })
+        
+        Toast.fire({
+          icon: 'success',
+          title: 'Gig deleted'
+        })
+      if(response.deleteGig===true){
+        this.getAllgigs()
+      }
+    })
+
     }
   })
+
+
+ 
 }
 
 }
